define types before hand
