+++ .gitignore
@@ -3,7 +3,6 @@ node_modules
 .pnpm-*
 .magefront-*
 dist
-types
 _*.md
 .turbo
 .secrets
