+++ packages/magefront/src/actions/watch.ts
@@ -2,7 +2,7 @@ import chokidar from 'chokidar'
 import path from 'node:path'
 import { performance } from 'node:perf_hooks'
 import prettyMilliseconds from 'pretty-ms'
-import { Action } from '../types'
+import type { Action } from '../../types/magefront'
 import { instance } from './browser-sync'
 import { build } from './build'
 import { deploy } from './deploy'
