+++ packages/magefront/src/main.ts
@@ -1,9 +1,7 @@
 // Main module entry point if the package is installed as dependency.
 // Gives quick access to the magento related functions from the package.
 
-import type { Config } from './types'
-
-export type { Config, MagefrontOptions, MagentoLanguage, MagentoModule, MagentoTheme, Plugin, PluginContext, Preset } from './types'
+import type { Config } from '../types/magefront'
 
 export { magefront } from './magefront'
