+++ packages/magefront/tests/theme-adminhtml-backend.test.ts
@@ -4,7 +4,7 @@ import { build } from '../src/actions/build'
 import { clean } from '../src/actions/clean'
 import { deploy } from '../src/actions/deploy'
 import { inheritance } from '../src/actions/inheritance'
-import { ActionContext } from '../src/types'
+import type { ActionContext } from '../types/magefront'
 import { testActionContext } from './helpers'
 
 describe('Build and deploy the Magento/backend theme', () => {
